issues resolved

+++ order/internal/service/order/delete.go
@@ -3,20 +3,26 @@ package order
 import (
 	"context"
 	"log"
+
+	"github.com/agumiroff/BigTechProject/shared/apperrors"
 )
 
 func (s *service) DeleteOrder(ctx context.Context, uuid string) error {
+	if uuid == "" {
+		return apperrors.ErrInvalidRequest
+	}
+
 	order, err := s.Repo.Get(ctx, uuid)
 	if err != nil {
-		log.Printf("failed to find order: %v", err)
+		log.Printf("Order with uuid does not exist %v: %v", uuid, err)
 		return err
 	}
 
 	err = s.Repo.DeleteOrder(ctx, order.OrderUUID)
 	if err != nil {
-		log.Printf("failed to delete order: %v", err)
 		return err
 	}
+	log.Printf("Order deleted")
 
 	return nil
 }
